Move model objects to internal package
IdempotentBlock
ConstructorBlock
TaggingFilterCriteria
