+++ src/main/java/com/greghaskins/spectrum/internal/IdempotentBlock.java
@@ -1,4 +1,4 @@
-package com.greghaskins.spectrum.model;
+package com.greghaskins.spectrum.internal;
 
 import com.greghaskins.spectrum.Block;
