+++ src/main/java/com/greghaskins/spectrum/internal/Suite.java
@@ -2,7 +2,6 @@ package com.greghaskins.spectrum.internal;
 
 import com.greghaskins.spectrum.Block;
 import com.greghaskins.spectrum.model.BlockConfiguration;
-import com.greghaskins.spectrum.model.TaggingFilterCriteria;
 
 import org.junit.runner.Description;
 import org.junit.runner.notification.RunNotifier;
