+++ src/main/java/com/greghaskins/spectrum/internal/junit/RuleContext.java
@@ -3,8 +3,8 @@ package com.greghaskins.spectrum.internal.junit;
 import static com.greghaskins.spectrum.internal.junit.StubJUnitFrameworkMethod.stubFrameworkMethod;
 
 import com.greghaskins.spectrum.Block;
+import com.greghaskins.spectrum.internal.ConstructorBlock;
 import com.greghaskins.spectrum.internal.Hook;
-import com.greghaskins.spectrum.model.ConstructorBlock;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
